requirements.txt
Flask==2.3.3
boto3==1.35.30
python-dotenv==1.0.1
